\name{alkdata}
\alias{alkdata}
\docType{data}
\title{Age-Length Key for Gulf of Hauraki snapper, 1992-1993}
\description{
The \code{alkdata} data frame has 39 rows and 16 columns.
The age-length key for Gulf of Hauraki snapper shown in Table 8.3 of Quinn and Deriso (1999)}
\format{
This data frame contains the following columns:
\describe{
\item{len}{length interval}
\item{nl}{number measured in length interval}
\item{A3}{number of fish aged in each age class 3 within each length interval} 
\item{A4}{number of fish aged in each age class 4 within each length interval} 
\item{A5}{number of fish aged in each age class 5 within each length interval} 
\item{A6}{number of fish aged in each age class 6 within each length interval} 
\item{A7}{number of fish aged in each age class 7 within each length interval} 
\item{A8}{number of fish aged in each age class 8 within each length interval} 
\item{A9}{number of fish aged in each age class 9 within each length interval} 
\item{A10}{number of fish aged in each age class 10 within each length interval} 
\item{A11}{number of fish aged in each age class 11 within each length interval} 
\item{A12}{number of fish aged in each age class 12 within each length interval} 
\item{A13}{number of fish aged in each age class 13 within each length interval} 
\item{A14}{number of fish aged in each age class 14 within each length interval} 
\item{A15}{number of fish aged in each age class 15 within each length interval} 
\item{A16}{number of fish aged in each age class 16 within each length interval} 
}
}
\source{
  Quinn, T. J. and R. B. Deriso. 1999. \emph{Quantitative Fish Dynamics}.
Oxford University Press, New York, NY. 542 p. 
}
\usage{
alkdata
}
\keyword{datasets}
