\name{AtkaMack}
\alias{AtkaMack}
\docType{data}
\title{Length and Age Data For Male and Female Atka Mackerel}
\description{
The \code{AtkaMack} data frame has 20 rows and 4 columns.
Mean length-at-age data for male and female Atka Mackerel as listed in Table 3 of Kimura (1990) }
\format{
This data frame contains the following columns:
\describe{
\item{age}{fish age}
\item{len}{mean length of fish of \emph{age} (cm)}
\item{sex}{sex code}
\item{m}{transformed age for SFR parameterization of von Bertalanffy equation}
\item{n}{sample size}
 
}
}
\source{
 Kimura, D. K. 1990. \emph{Testing nonlinear regression paramters under heteroscedastic, normally distributed errors}. 
Biometrics 46:697-708.
}
\usage{
AtkaMack
}
\keyword{datasets}
