\name{buffalo}
\alias{buffalo}
\docType{data}
\title{Life Table Data for African Buffalo}
\description{
The \code{buffalo} data frame has 20 rows and 3 columns.
Cohort size and deaths for African buffalo from Sinclair (1977) as reported by Krebs (1989) in
Table 12.1, page 415. 
}
\format{
This data frame contains the following columns:
\describe{
\item{age}{age interval}
\item{nx}{ number alive at start of each age interval}
\item{dx}{number dying between age interval X and X+1}
}
}
\source{
 Krebs, C. J. 1989. \emph{Ecological Methodologies}. Harper and Row, New York, NY. 654 p.}
\usage{
buffalo
}
\keyword{datasets}
