
\name{catch}
\alias{catch}
\docType{data}
\title{Number of cod captured in 10 standardized bottom trawl hauls from Massachusetts, 1985}
\description{
The \code{catch} data frame has 10 rows and 1 column.
 }
\format{
This data frame contains the following columns:
\describe{
\item{value}{catch data}
}
}
\source{Massachusetts Division of Marine Fisheries}
\usage{
catch
}
\keyword{datasets}
