\name{codcluslen}
\alias{codcluslen}
\docType{data}
\title{Lengths of Atlantic cod caught during Massachusetts Division of Marine Fisheries bottom trawl survey, spring 1985.}
\description{
The \code{codcluslen} data frame has 334 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{region}{NorthCape = North of Cape Cod; SouthCape =South of Cape Cod}
\item{tow}{Tow number}
\item{length}{Length class (total length, cm)} 
\item{number}{Number in length class} 
}
}
\source{Massachusetts Division of Marine Fisheries}
\usage{
codcluslen
}
\keyword{datasets}
