\name{codlengths}
\alias{codlengths}
\docType{data}
\title{Individual lengths of Atlantic cod from cooperative trawling in winter of 2000, Massachusetts Bay}
\description{
The \code{codlengths} data frame has 184 rows and 3 columns.
Station, total number of fish caught, and total lengths (cm) of Atlantic cod (\emph{Gadus morhua}) collected via trawling from northern Massachusetts
during winter of 2000}
\format{
This data frame contains the following columns:
\describe{
\item{station}{trawl haul ID}
\item{total}{total number of cod caught in trawl haul}
\item{tl}{total length of individual fish measured in each haul}

}
}
\source{
 Massachusetts Division of Marine Fisheries, 30 Emerson Avenue, Gloucester, MA 01930.
}
\usage{
codlengths
}
\keyword{datasets}
