\name{codstrcluslen}
\alias{codstrcluslen}
\docType{data}
\title{Lengths of Atlantic cod caught during Massachusetts Division of Marine Fisheries stratified random bottom trawl survey, spring 1985.}
\description{
The \code{codstrcluslen} data frame has 334 rows and 6 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{region}{NorthCape = North of Cape Cod; SouthCape = South of Cape Cod}
\item{stratum}{Stratum number}
\item{tow}{Tow number}
\item{weights}{Stratum area (square nautical-miles)}
\item{length}{Length class (total length cm)} 
\item{number}{Number in length class}
}
}
\source{Massachusetts Division of Marine Fisheries, 30 Emerson Avenue, Gloucester, MA 01930
}
\usage{
codstrcluslen
}
\keyword{datasets}
