\name{counts}
\alias{counts}
\docType{data}
\title{Run size data for alewife (\emph{Alosa pseudoharengus})}
\description{
The \code{counts} data frame has 31 rows and 2 columns.
 Run size data of alewife (\emph{Alosa pseudoharengus}) in Herring River, MA from 1980-2010
}
\format{
This data frame contains the following columns:
\describe{
\item{year}{vector of run year}
\item{number}{vector of run counts in number of fish}
}
}
\source{
 Massachusetts Division of Marine Fisheries, 30 Emerson Avenue, Gloucester, MA
}
\usage{
counts
}
\keyword{datasets}
