\name{cowcod}
\alias{cowcod}
\docType{data}
\title{Catch data (metric tons) for cowcod Sebastes levis 1900 to 2008}
\description{
Cowcod catch data from literature sources in Martell and Froese (2012).  
}
\format{
  A data frame with 109 observations on the following 2 variables.
  \describe{
    \item{\code{year}}{a numeric vector describing the year of catch}
    \item{\code{catch}}{a numeric vector describing the annual catch in metric tons}
   }
}
\usage{cowcod}
\keyword{datasets}

