
\name{darter}
\alias{darter}
\docType{data}
\title{Catch Removal Data For Fantail Darter}
\description{
The \code{darter} data frame has 7 rows and 2 columns.
Sequence of catch data for the faintail darter from removal experiments by Mahon as reported by
White et al.(1982).  This dataset is often use to test new depletion estimators because the actual abundance is known (N=1151).
 }
\format{
This data frame contains the following columns:
\describe{
\item{catch}{catch data}
\item{effort}{constant effort data}
}
}
\source{
White, G. C., D. R. Anderson, K. P. Burnham, and D. L. Otis. 1982. \emph{Capture-recapture and Removal Methods for Sampling
Closed Populations}. Los Alamos National Laboratory LA-8787-NERP. 235 p.
}
\usage{
darter
}
\keyword{datasets}
