
\name{ep.data}
\alias{ep.data}
\docType{data}
\title{A sub-sample of data from a simulated population collected via length-stratified age sampling}
\description{
The \code{catch} data frame has 1072 rows and 4 columns.
 }
\format{
This data frame contains the following columns:
\describe{
\item{length}{length in cm.}
\item{age}{age of fish (yrs).}
\item{Nh}{the total sample size per bin (includes unaged fish).}
\item{nh}{the total aged sample size per bin.}

}
}
\source{Andrea Perrault,  Marine Institute of Memorial University of Newfoundland}
\usage{
ep.data
}
\keyword{datasets}
