\name{fishmethods-package}
\alias{fishmethods-package}
\alias{fishmethods}
\docType{package}
\title{
Fisheries Methods and Models in R
}
\description{
Fisheries methods and models that extract Marine Recreational Fisheries
 Statistics data, estimate population parameters (e.g., mortality, abundance,
 age structure, etc.), run dynamic pool models, and compare population
 characteristics using special statistical techniques required in fisheries
 sampling.
}
\details{
\tabular{ll}{
Package: \tab fishmethods\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2008-12-30\cr
License: \tab GPL(>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Gary A. Nelson, Massachusetts Division of Marine Fisheries, Gloucester, MA.

Maintainer: Gary A. Nelson <Gary.Nelson@state.ma.us>
}
\references{
Specified for each function under help pages.
}
\keyword{ package }
