\name{Gerking}
\alias{Gerking}
\docType{data}
\title{Mark-Recapture Data for Sunfish in an Indiana Lake}
\description{
The \code{Gerking} data frame has 14 rows and 3 columns.
Marked and released sunfish in an Indiana lake for 14 days by Gerking (1953) as reported by Krebs (1989, Table 2.1).
}
\format{
This data frame contains the following columns:
\describe{
\item{C}{column of number of captures (column names is unnecessary).}
\item{R}{column of number of recaptures (column name is unnecessary).}
\item{nM}{column of number of newly marked animal (column name is unnecessary).}
}
}
\source{
 Krebs, C. J. 1989. \emph{Ecological Methodologies}. Harper and Row, New York, NY. 654 p.}
\usage{
Gerking
}
\keyword{datasets}
