\name{goosefish}
\alias{goosefish}
\docType{data}
\title{Mean Length and Numbers of Lengths for Northern Goosefish, 1963-2002}
\description{
The \code{goosefish} data frame has 40 rows and 3 columns.
The mean lengths (mlen) by year and number (ss) of observations for length>=smallest length at first capture (Lc)
for northern goosefish used in Gedamke and Hoenig (2006)
}
\format{
This data frame contains the following columns:
\describe{
\item{year}{year code}
\item{mlen}{mean length of goosefish, total length (cm)}
\item{ss}{number of samples used to calculate mean length}
}
}
\source{Gedamke, T. and J. M. Hoenig. 2006. Estimating mortality from mean length
  data in nonequilibrium situations, with application to the assessment of goosefish. Trans. Am. Fish. Soc. 135:476-487}
\usage{
goosefish
}
\keyword{datasets}
