
\name{haddock}
\alias{haddock}
\docType{data}
\title{Biological data for haddock (Melanogrammus aeglefinus)}
\description{
The \code{haddock} data frame has 15 rows and 4 columns.
Age, weight at spawning, partial recruitment, and fraction mature data for haddock (Melanogrammus aeglefinus) used by Gabriel et al. (1989) 
to calculate spawning stock biomass-per-recruit.
}
\format{
This data frame contains the following columns:
\describe{
\item{age}{vector of ages}
\item{ssbwgt}{vector of weights at spawning for each age}
\item{partial}{partial recruitment vector}
\item{pmat}{vector of fraction of females mature at age}

}
}
\source{
 Gabriel, W. L., M. P. Sissenwine, and W. J. Overholtz. 1989. Analysis of spawning stock biomass per recruit:
an example for Georges Bank haddock. North American Journal of Fisheries Management 9: 383-391.
}
\usage{
haddock
}
\keyword{datasets}
