
\name{herring}
\alias{herring}
\docType{data}
\title{Total length data for alewive (Alosa pseudoharengus) from Massachusetts waters}
\description{
The \code{herring} data frame has 52 rows and 1 column.
Total length (mm)
}
\format{
This data frame contains the following columns:
\describe{
\item{tl}{total length of individual alewife}
}
}
\source{
 Massachusetts Division of Marine Fisheries, 30 Emerson Avenue, Gloucester, MA 01930

}
\usage{
herring
}
\keyword{datasets}
