\name{Hightower}
\alias{Hightower}
\docType{data}
\title{Original data used in Hightower et al. (2001)}
\description{
The \code{Hightower} has 51 rows and 1 column.
The complete capture histories of striped bass for Lake Gaston, North Carolina. 
}
\format{
This data frame contains the following columns:
\describe{
\item{caphistory}{capture histories of 51 striped bass}
}
}
\source{
Hightower, J. E., J. R. Jackson, and K. H. Pollock. 2001. Use of telemetry methods to estimate natural mortality and fishing mortality of striped bass in Lake Gaston, North Carolina. Trans. Am. Fish. Soc. 130:557-567.

Thanks to Joe Hightower of NC Cooperative Fish and Wildlife Research Unit for providing his original data.
}
\usage{
Hightower
}
\keyword{datasets}
