\name{Hoenig}
\alias{Hoenig}
\docType{data}
\title{Tag Data from Hoenig et al. (1998)}
\description{
The \code{Hoenig} list containing 8 components of data.
Data were obtained from the Hoenig et al.(1998).
}
\format{
This list contains the following components:
\describe{
\item{relyrs}{vector of start and end years of release years}
\item{recapyrs}{vector of start and end years of recapture years}
\item{N}{vector of number of tags released in each release year}
\item{recapharv}{recapture matrix of harvested fish}
\item{lambda}{vector of reporting rates (one for each recapture year)}
\item{phi}{vector of initial tag loss (one for each recapture year)}
\item{Fyr}{vector of years to estimate fishing mortality}
\item{Myr}{vector of years to estimate natural mortality}
}
}
\source{
  Hoenig, J. M, N. J. Barrowman, W. S. Hearn, and K. H. Pollock. 1998. Multiyear tagging studies incorporating fishing effort data.  Canadian Journal of Fisheries and Aquatic Sciences 55: 1466-1476.
}
\usage{
Hoenig
}
\keyword{datasets}
