\name{hohe}
\alias{hohe}
\docType{data}
\title{age-length key and length frequency data from Hoenig and Heisey (1987)}
\description{
The \code{hohe} data list with age-length key matrix with 10 columns and 4 rows, and length frequency vector with 10 observations .
Age-length key and length frequency from Appendix B example in Hoeing and Heisey (1987)}
\format{
 One matrix and one vector
}
\source{
 Hoenig, J. M. and D. M. Heisey. 1987. \emph{Use of a log-linear model with the EM algorithm to correct
estimates of stock composition and to convert length to age. Transactions of the American Fisheries Society 116: 232-243.}
}
\usage{
hohe
}
\keyword{datasets}
