
\name{Jensen}
\alias{Jensen}
\docType{data}
\title{Age Frequency Data for Lake Whitefish By Individual Haul}
\description{
The \code{Jensen} data frame has 312 rows and 2 columns.
The age data are from reconstructed catches of lake whitefish reported
by Jensen (1996) in Table 1 and were expanded to individual observations from the age frequency table. 
}
\format{
This data frame contains the following columns:
\describe{
\item{group}{net haul label}
\item{age}{age of an individual fish}
}
}
\source{
  Jensen, A. L. 1996. \emph{Ratio estimation of mortality using catch curves}. Fisheries Research 27: 61-67.
}
\usage{
Jensen
}
\keyword{datasets}
