\name{Jiang}
\alias{Jiang}
\docType{data}
\title{Tag Data from Jiang (2005)}
\description{
The \code{Jiang} list containing 13 components of data.
Data were obtained from the Jiang (2005).
}
\format{
This list contains the following components:
\describe{
\item{relyrs}{vector of start and end years of release years}
\item{recapyrs}{vector of start and end years of recapture years}
\item{N}{vector of number of tags released in each release year}
\item{recapharv}{recapture matrix of harvest fish}
\item{recaprel}{recapture matrix of recaptured and re-released fish with tag removed}
\item{hlambda}{vector of reporting rates of harvested fish (one value for each recapture year)}
\item{rlambda}{vector of reporting rates of recaptured and re-released fish (one value for each recapture year)}
\item{hphi}{vector of initial tag loss of harvested fish (one value for each recapture year)}
\item{rphi}{vector of initial tag loss of harvested fish (one value for each recapture year)}
\item{hmrate}{vector of hooking mortality rates (one value for each recapture year)}
\item{Fyr}{vector of years to estimate fishing mortality}
\item{FAyr}{vector of years to estimate tag mortality}
\item{Myr}{vector of years to estimate natural mortality}
}
}
\source{
  Jiang, H. 2005. Age-dependent tag return models for estimating fishing mortality, natural mortality and selectivity. 
Doctoral dissertation. North Carolina State  University, Raleigh.

}
\usage{
Jiang
}
\keyword{datasets}
