\name{kappenman}
\alias{kappenman}
\docType{data}
\title{Pacific cod catch per effort from Table 1 in Kappenman (1999)}
\description{
The \code{kappenman} data frame has 55 rows and 1 column.}
\format{
This data frame contains one column:
\describe{
\item{cpue}{Pacific cod cpue from 1994}
}
}
\source{
 Kappenman, R. F. 1999. Trawl survey based abundance estimation using data sets 
with unusually large catches. ICES Journal of Marince Science 56: 28-35.
}
\usage{
kappenman
}
\keyword{datasets}
