
\name{Kimura}
\alias{Kimura}
\docType{data}
\title{Length and Age Data For Male and Female Pacific Hake}
\description{
The \code{Kimura} data frame has 24 rows and 3 columns.
Mean length-at-age data for male and female Pacific hake as reported by Kimura (1980)}
\format{
This data frame contains the following columns:
\describe{
\item{age}{fish age}
\item{length}{mean length of fish of age \emph{age}}
\item{sex}{sex code} 
}
}
\source{
 Kimura, D. K. 1980. \emph{Likelihood methods for the von Bertalanffy growth curve}. 
U. S. Fishery Bulletin 77:765-776.
}
\usage{
Kimura
}
\keyword{datasets}
