\name{lepdata}
\alias{lepdata}
\docType{data}
\title{Simulated data based on parameters estimated from corrected 1980s bluefin tuna data used in Laslett et al.(2002)}
\description{
The \code{lepdata} data frame has 500 rows and 3 columns
}
\format{
This data frame contains the following columns:
\describe{
\item{l1}{release length}
\item{l2}{reapture length}
\item{dt}{time increment between release and recapture}
}
}
\source{
 Original data provided by Paige Eveson of CSIRO Marine Research.
 }
\usage{
lepdata
}
\keyword{datasets}
