\name{lingcod}
\alias{lingcod}
\docType{data}
\title{Catch data (metric tons) for lingcod 1889 to 2001}
\description{
Lingcod catch data from literature sources in Martell and Froese (2012).  
}
\format{
  A data frame with 113 observations on the following 2 variables.
  \describe{
    \item{\code{year}}{a numeric vector describing the year of catch}
    \item{\code{catch}}{a numeric vector describing the annual catch in metric tons}
   }
}
\details{
Note some data points are not exactly the same as shown in Figure 7 of Martell and Froese 2012.
}
\usage{lingcod}
\keyword{datasets}

