\name{maki}
\alias{maki}
\docType{data}
\title{Data from Maki et al. 2001}
\description{
The \code{maki} data frame has 876 rows and 2 columns.
 From Table 1 for 3 years combined
}
\format{
This data frame contains the following columns:
\describe{
\item{capture_age}{age at capture}
\item{age_mature}{age at first maturity (from spawning checks on scales)}
}
}
\source{
 Maki, K. L., J. M. Hoenig and J. E. Olney. 2001. Estimating proportion mature at age 
when immature fish are unavailable for study, with applications to American shad in the York
River, Virginia. North Am. J. Fish. Manage. 21: 703-716.}
\usage{
maki
}
\keyword{datasets}
