
\name{menhaden}
\alias{menhaden}
\docType{data}
\title{Biological data for menhaden (Brevoortia tyrannus)}
\description{
The \code{menhaden} data frame has 15 rows and 4 columns.
Age, fecundity-at-age, partial recruitment, fraction mature, and nautral mortality data for menhaden 
to calculate eggs-per-recruit.
}
\format{
This data frame contains the following columns:
\describe{
\item{age}{vector of ages}
\item{fecundity}{vector of mean eggs per individual for each age}
\item{partial}{partial recruitment vector}
\item{pmat}{vector of fraction of females mature at age}
\item{M}{vector of natural mortality value-at-age}
}
}
\source{
  Atlantic State Marine Fisheries Commission. 2010. 2009 stock assessment report for Atlantic menhaden. ASMFC SAR 10-02.
 }
\usage{
menhaden
}
\keyword{datasets}
