\name{nshrimp}
\alias{nshrimp}
\alias{nshrimp}

\docType{data}
\title{Data for Gulf of Maine northern shrimp}
\description{
Recruit and postrecruit survey indices and catch data for Gulf of Maine northern shrimp (Pandulus borealis), 1985-2007}
\usage{data(nshrimp)}
\format{
  A data frame with 23 observations on the following 4 variables.
  \describe{
    \item{\code{year}}{a numeric vector describing the year}
    \item{\code{r}}{a numeric vector of the recruit index}
    \item{\code{n}}{a numeric vector of the postrecruit index}
    \item{\code{C}}{a numeric vector of the landings (in numbers)}
  }
}
\source{
\url{http://www.nefsc.noaa.gov/nefsc/publications/crd/crd0716/a.pdf}
}
\keyword{datasets}

