\name{P.donacina}
\alias{P.donacina}
\docType{data}
\title{Data from a growth study of New Zealand intertidal clams.}
\description{
Growth increment data derived from a tagging experiment on Paphis donacina
}
\format{
  A data frame with 150 observations on the following 4 variables.
  \describe{
    \item{\code{T1}}{a numeric vector describing the release date (y)}
    \item{\code{T2}}{a numeric vector describing the recovery date (y)}
    \item{\code{L1}}{a numeric vector describing the length at release (mm)}
    \item{\code{L2}}{a numeric vector describing the length at recapture (mm)}
  }
}
\details{Note that the data have been corrected for measurement bias, as
described by Cranfield et al (1996).}

\source{Cranfield, H.J., Michael, K.P., and Francis, R.I.C.C. 1996.
  Growth rates of five species of subtidal clam on a beach in the South
  Island, New Zealand.  Marine and Freshwater Research 47: 773--784.}
\usage{P.donacina}

\keyword{datasets}

