\name{pinfish}
\alias{pinfish}
\docType{data}
\title{Length, age and sex data for pinfish (Lagodon rhomboides) from Tampa Bay, Florida}
\description{
The \code{pinfish} data frame has 670 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{field_no}{haul identifier}
\item{sl}{standard length (mm) of individual pinfish}
\item{age}{age in year with decimal extention reflecting age past January 1}
\item{sex}{sex of fish. 1=male, 2=female, 0 = unknown}
}
}
\source{
Nelson, G. A.  2002. Age, growth, mortality, and distribution of pinfish (Lagodon rhomboides)
 in Tampa Bay and adjacent Gulf of Mexico waters.  Fishery Bulletin 100: 582-592.
}
\usage{
pinfish
}
\keyword{datasets}
