\name{rig}
\alias{rig}
\docType{data}
\title{Tagging data from a growth study of rig}
\description{
Tagging growth increment data for New Zealand rig (Mustelus
lenticulatus), after removal of outliers, as analysed in models 2-4 
of Table 6 of Francis and Francis (1992).
}
\format{
  A data frame with 114 observations and the following components
  \describe{
    \item{\code{L1}}{Length at release (cm)}
    \item{\code{L2}}{Length at recapture (cm)}
    \item{\code{T1}}{Time of release (y from 1 January 1981)}
    \item{\code{T2}}{Time of recapture (y from 1 January 1981)}
    \item{\code{Sex}}{Sex of fish (F or M)}
  }
}
\source{
\describe{\item{1}{Francis, M.P. and Francis, R.I.C.C. 1992.  Growth rate
  estimates for New Zealand rig (Mustelus lenticulatus).  Australian
  Journal of Marine and Freshwater Research 43: 1157--1176}
}
}
\usage{rig}

\keyword{datasets}

