\name{rockbass}
\alias{rockbass}
\docType{data}
\title{Age Frequency Data for Rock Bass}
\description{
The \code{rockbass} data frame has 243 rows and 1 column.
The age data are from a sample of rock bass trap-netted from Cayuga Lake, New York by Chapman and Robson, as reported
by Seber (2002; page 417) and were expanded to individual observations from the age frequency table. 
}
\format{
This data frame contains the following columns:
\describe{
\item{age}{age of individual rock bass in years}
}
}
\source{
  Seber, G. A. F. 2002. \emph{The Estimation of Animal Abundance and Related Parameters, Second Edition}.
The Blackburn Press, Caldwell, New Jersey. 654 p.
}
\usage{
rockbass
}
\keyword{datasets}
