\name{rockfish}
\alias{rockfish}
\docType{data}
\title{Individual length and age data for rougheye rockfish}
\description{
The \code{rockfish} data frame has 545 rows and 2 columns.
The individual length-at-age data were generated from mean length-at-age data and standard deviations 
   given in Table 4.3 of Quinn and Deriso (1990). 
}
\format{
This data frame contains the following columns:
\describe{
\item{age}{age of individual rockfish in years}
\item{len}{length of individual rockfish in cm}

}
}
\source{
  Quinn, T. J. and R. B. Deriso. 1999. Quantitative fish dynamics. Oxford University Press. 542 pages. 
}
\usage{
rockfish
}
\keyword{datasets}
