\name{sblen}
\alias{sblen}
\docType{data}
\title{Total length (inches) of striped bass collected by Massachusetts volunteer anglers in 2014}
\description{
\code{sblen} data frame has 311 rows and 1 columns.
 Total length of striped bass
}
\format{
This data frame contains the following columns:
\describe{
\item{len_inches}{vector of lengths}
}
}
\source{
 Massachusetts Division of Marine Fisheries, 30 Emerson Avenue, Gloucester, MA
}
\usage{
sblen
}
\keyword{datasets}
