\name{sbotos}
\alias{sbotos}
\docType{data}
\title{Otolith ages of striped bass made by two age readers}
\description{
The \code{sbotos} data frame has 135 rows and 2 columns.
 Ages of striped bass interpreted from the same otolith sections by two age readers 
}
\format{
This data frame contains the following columns:
\describe{
\item{reader1}{vector of ages}
\item{reader2}{vector of ages}
}
}
\source{
 Massachusetts Division of Marine Fisheries, 30 Emerson Avenue, Gloucester, MA
}
\usage{
sbotos
}
\keyword{datasets}
