\name{Shepherd}
\alias{Shepherd}
\docType{data}
\title{Seasonal Length Frequencies for Raja clavata}
\description{
The \code{Shepherd} data frame has 24 rows and 4 columns.
The seasonal length frequency data of Raja clavata are from Shepherd's working document. 
}
\format{
This data frame contains the following columns:
\describe{
\item{length}{lower limit of length interval }
\item{f1}{length frequency from first sampling event in year.}
\item{f2}{length frequency from second sampling event in year.}
\item{f3}{length frequency from third sampling event in year.}
}
}
\source{Shepherd, J. G. 1987. \emph{A weakly parametric method for the analysis of length composition data}. In: D. Pauly and
G. Morgan, (eds). The Theory and Application of Length-Based Methods of Stock Assessment. ICLARM Conf. Ser. Manilla.}
\usage{
Shepherd
}
\keyword{datasets}
