\name{sole}
\alias{sole}
\alias{sole}
\docType{data}
\title{Flathead sole CPUEs}
\description{
Flathead sole CPUEs for a side-by-side trawl calibration study of National Marine Fisheries Service (NMFS) and Alaska Department of Fish and Game (ADFG) vessels}
\usage{data(sole)}
\format{
  A data frame with 33 observations on the following 3 variables.
  \describe{
    \item{\code{haul}}{a numeric vector of the experimental paired haul number}
    \item{\code{nmfs}}{catch-per-unit-effort (kg per km2) for the NMFS vessel Peggy Jo from 33 experimental hauls}
    \item{\code{adfg}}{catch-per-unit-effort (kg per km2) for the ADFG vessel Resolution from 33 experimental hauls}
  }
}
\source{

von Szalay, P. G. and E. Brown. 2001. Trawl comparisons of fishing power differences and
their applicability to National Marine Fisheries Service and Alask Department of Fish
and Game trawl survey gear.  Alaska Fishery Research Bulletin 8(2):85-95.

Data were graciously provided by Paul G. von Szalay, National Marine Fisheries Service, Seattle, Washington.

}
\keyword{datasets}

