\name{striper}
\alias{striper}
\docType{data}
\title{Recruitment Numbers and Female Spawning Stock Biomass for Striped Bass}
\description{
The \code{striper} data frame has 34 rows and 2 column.
Estimates of recruits and female spawning stock biomass for striped bass from 
the Atlantic State
Marine Fisheries 2016 stock assessment.
}
\format{
This data frame contains the following columns:
\describe{
\item{recruits}{number of recruits}
\item{stock}{female spawning stock biomass (metric tons)}
}
}
\source{
 \url{http://www.asmfc.org}
}
\usage{
striper
}
\keyword{datasets}
