\name{tanaka}
\alias{tanaka}
\docType{data}
\title{Simulated alfonsino data for Tanaka (2006}
\description{
The \code{tanaka} data frame has  138 rows and 3 columns.
The number of returns and the mean times-at-large from Table 2 of 
Tanaka (2006) were used to generate individual times-at-large data from a random normal distributions
using a CV of 0.1. 
}
\format{
This data frame contains the following columns:
\describe{
\item{relyr}{release year (cohort)}
\item{tal}{individual times-at-large (in years)}
\item{N}{Total number of releases for relear year (cohort)}
}
}
\source{
Tanaka, E. 2006. Simultaneous estimation of instantaneous mortality coefficients and rate 
of effective survivors to number of released fish using multiple sets of tagging experiments. 
Fisheries Science 72: 710-718.
}
\usage{
tanaka
}
\keyword{datasets}
