
\name{trout}
\alias{trout}
\docType{data}
\title{Mark-recapture data for Kenai River trout trout}
\description{
The \code{trout} data frame has 102 rows and 3 columns.
Release lengths, recapture lengths and times-at-large for trout trout in the Kenai River from Table 4.10 of Quinn and Deriso (1999).
}
\format{
This data frame contains the following columns:
\describe{
\item{L1}{vector of release lengths}
\item{L2}{vector of recapture lengths}
\item{dt}{vector of times-at-large }
}
}
\source{
Quinn, T. J. and R. B. Deriso. 1999. Quantitative Fish Dynamics. Oxford University Press, New York, New York. 542 pages
}
\usage{
trout
}
\keyword{datasets}
