\name{wolffish}
\alias{wolffish}
\docType{data}
\title{Spring untransformed mean catch per tow for wolffish (Anarhichas lupus)}
\description{
The \code{wolffish} data frame has 25 rows and 2 columns.
The  mean catch per tow values were digitized from Figure 4 of Helser and Hayes (1995) and back-transformed to the original scale.
}
\format{
This data frame contains the following columns:
\describe{
\item{year}{survey year of catch per tow}
\item{index}{mean catch per tow value (untransformed)}
}
}
\source{
  Helser, T. E. and D. B. Hayes. 1995.  Providing quantitative management advice from stock abundance 
indices based on research surveys. Fishery Bulletin 93: 290-298.}
\usage{
wolffish
}
\keyword{datasets}
