\name{yellowtail}
\alias{yellowtail}
\docType{data}
\title{Fall average catch per tow for southern New England yellowtail flounder}
\description{
The \code{yellowtail} data frame has 22 rows and 2 columns.
The average catch per tow values were digitized from Figure 4 of Pennington (1986)
}
\format{
This data frame contains the following columns:
\describe{
\item{year}{survey year of catch per tow}
\item{index}{average catch per tow value (untransformed)}
}
}
\source{
  Pennington, M. P. 1986. Some statistical techniques for estimating abundance indices from trawl surveys. Fishery Bulletin 84(3): 519-525.
}
\usage{
yellowtail
}
\keyword{datasets}
