\name{bonito}
\alias{bonito}
\docType{data}
\title{Data from an age and growth study of the pacific bonito.}
\description{
Growth increment data derived from tagging experiments on Pacific bonito
(Sarda chiliensis) used to illustrate Francis's maximum likelihood
method estimation of growth and growth variability (1988).
}
\format{
  A data frame with 138 observations on the following 4 variables.
  \describe{
    \item{\code{T1}}{a numeric vector describing the release date}
    \item{\code{T2}}{a numeric vector describing the recovery date}
    \item{\code{L1}}{a numeric vector describing the length at release
      in cenitmeters}
    \item{\code{L2}}{a numeric vector describing the length at recapture
      in centimeters}
  }
}
\details{
Note that Francis (1988) has discarded several records from the original
dataset collected by Campbell et al. (1975).
}

\source{
\describe{\item{1}{Francis, R.I.C.C., 1988. Maximum likelihood estimation of growth and growth variability from tagging data. New Zealand Journal of Marine and Freshwater Research, 22, p.42--51.}
\item{2}{Campbell, G. & Collins, R., 1975. The age and growth of the Pacific bonito, Sarda chiliensis, in the eastern north Pacific. Calif. Dept. Fish Game, 61(4), p.181-200.}
}
}
\usage{bonito}

\keyword{datasets}

