\name{striper}
\alias{striper}
\docType{data}
\title{Age and length data of the striped bass}
\description{
Length and age (from scales) data for striped bass (\emph{Morone saxatilis}) 
from recreational and commercial fisheries in Massachusetts during 2009.
}
\format{
  A data frame with 846 observations on the following 2 variables.
  \describe{
    \item{\code{age}}{a numeric vector of individual ages}
    \item{\code{tl}}{a numeric vector of individual lengths (total length in millimeter)}
  }
}

\source{Massachusetts Division of Marine Fisheries
}

\usage{striper}
\keyword{datasets}

